pyproject.toml
[project]
name = "pt-appsec-ml-joleal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = []

[dependency-groups]
dev = [
    "python-dotenv>=1.1.1",
    "strands-agents-tools>=0.2.8",
    "strands-agents[ollama,openai]>=1.9.1",
]
notebook = [
    "jupyter>=1.1.1",
]
